requirements.txt
asgiref==3.11.0
django==6.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
pyjwt==2.10.1
sqlparse==0.5.5
